add: DELETE for encrypted clipboards
